% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpdf.R
\name{ggpdf}
\alias{ggpdf}
\title{Save ggplot figures as PDFs with good defaults}
\usage{
ggpdf(filename, plot, height, width, keep_svg = FALSE, ...)
}
\arguments{
\item{filename}{File name to create on disk.}

\item{plot}{Plot to save.}

\item{height}{Plot size in units ("in", "cm", "mm", or "px").}

\item{width}{Plot size in units ("in", "cm", "mm", or "px").}

\item{keep_svg}{Keep the svg file? Defaults to FALSE.}

\item{...}{Other arguments to pass to svg().}
}
\value{

}
\description{
Saves a ggplot figure to PDF. Importantly, this is done via SVG so that
text is converted to a static image. Optionally the svg can be retained.
}
