% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speed_conversions.R
\name{kmh_to_ms}
\alias{kmh_to_ms}
\alias{ms_to_kmh}
\alias{kmh_to_mph}
\alias{mph_to_kmh}
\alias{mph_to_ms}
\alias{ms_to_mph}
\title{Speed Conversions}
\usage{
kmh_to_ms(x)

ms_to_kmh(x)

kmh_to_mph(x)

mph_to_kmh(x)

mph_to_ms(x)

ms_to_mph(x)
}
\arguments{
\item{x}{Input vector.}
}
\description{
Convert between different vehicle speed units.
}
\author{
Jack Davison
}
